+++ Changelog.md
@@ -10,6 +10,8 @@ Compiler Features:
 
 
 Bugfixes:
+ * Control Flow Graph: Take internal calls to functions that always revert into account for reporting unused or unassigned variables.
+
 
 
 ### 0.8.4 (2021-04-21)